lifted state to app.js, passing data to result component from searchresults state
